pyproject.toml
[tool.poetry]
name = "tfg-django-service"
version = "0.0.0"
description = ""
authors = ["carlosruperez <carlosruperez.h@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
poetry-core = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
